requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.0
langchain>=0.3.7
langchain-google-genai>=2.0.4
langchain-community>=0.3.5
langchain-qdrant>=0.2.0
langchain-huggingface>=1.0.0
langchain-text-splitters>=1.0.0
langchain-classic>=0.1.0
qdrant-client>=1.12.0
sentence-transformers>=3.3.0
pypdf>=5.1.0
jinja2>=3.1.4
